chore: rename subpackage and classes; add typehints
