+++ ra2ce/analysis/damages/damage_functions/damage_fraction_uniform.py
